Revert "优化代码"
This reverts commit 6401dfffebffe42137dd82ca4a1c1d3eb3f8da62.

+++ app/src/main/java/com/ddu/ui/helper/AvatarImageBehavior.java
@@ -104,8 +104,8 @@ public class AvatarImageBehavior extends CoordinatorLayout.Behavior<ImageView> {
             child.setY(mStartYPosition - distanceYToSubtract);
 
             CoordinatorLayout.LayoutParams lp = (CoordinatorLayout.LayoutParams) child.getLayoutParams();
-            lp.width = mStartHeight;
-            lp.height = mStartHeight;
+            lp.width = (int) (mStartHeight);
+            lp.height = (int) (mStartHeight);
             child.setLayoutParams(lp);
         }
         return true;
